drm/i915/gvt: enable cursor plane emulation
Add handler to intercept cursor plane operation and update h/w with value
from guest OS.

The watermark from the guest OS and the predefined ddb in the service OS
are set to hardware together. It's workable but has risk. The guest OS
driver has a set of ddb and watermark setting. The ddb topology might be
conflict with the predefined ddb in virtualization case. So it is better
to re-compute the watermark in service OS with the plane information
from guest OS and predefined ddb together in future.

v3: combine small patches to enable cursor plane
v2: remove magic number and add comment

Signed-off-by: Liu Xinyun <xinyun.liu@intel.com>