drm/i915/gvt: update cursor plane watermark with the value from guest OS
Here the watermark from guest OS and the predefined ddb in service OS
are set to hardware together. It's workable but not precise. guest OS
driver has a set of ddb and watermark setting. The ddb info is conflict
with the predefined ddb usually in virtualization case.  So it is better
to re-compute the watermark in service OS with the plane information
from guest OS and predefined ddb together.

Signed-off-by: Liu Xinyun <xinyun.liu@intel.com>