drm/i915/gvt: update cursor plane watermark to H/W
the watermark from the guest OS and the predefined ddb in the service OS
are set to hardware together. It's workable but not precise. the guest OS
driver has a set of ddb and watermark setting. The ddb info is conflict
with the predefined ddb in virtualization case usually. So it is better
to re-compute the watermark in service OS with the plane information
from guest OS and predefined ddb together in future.

Signed-off-by: Liu Xinyun <xinyun.liu@intel.com>